+++ bin/.gitignore
@@ -3,3 +3,4 @@
 /Main/
 /com/
 /helloFx/
+/controller/
