+++ src/controller/FormularioProjetoScreenController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.miprogramacao.gerenciadordetarefas.controller;
+package controller;
 
 import java.io.IOException;
 import java.net.URL;
@@ -19,15 +19,13 @@ import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuBar;
-import javafx.scene.control.MenuButton;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import javafx.scene.control.Alert.AlertType;
-import javafx.scene.control.Button;
-import javafx.scene.control.DatePicker;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
+import com.miprogramacao.gerenciadordetarefas.model.*;
 
 
 /**
@@ -35,26 +33,44 @@ import javafx.stage.Stage;
  * @author User
  */
 
-public class FormularioTarefaScreenController implements Initializable {
+public class FormularioProjetoScreenController implements Initializable {
 	
-	@FXML
+    @FXML
     private TextField txtTitulo;
 
     @FXML
     private TextArea txtDescricao;
+    
+    private MessageAlert msgAlert = new MessageAlert();
+    
+    private Projeto temp = new Projeto("temp");
 
     @FXML
-    private DatePicker txtValidade;
-
-    @FXML
-    private MenuButton statusMenuBTN;
-
-    @FXML
-    private Button addNovaTarefa;
-
-    @FXML
-    void salvarNovaTarefa(ActionEvent event) {
-
+    void addAlteracao(ActionEvent event) throws ArgumentoInvalidoException {
+    	
+    	String titulo = txtTitulo.getText();
+    	String descricao = txtDescricao.getText();
+    	
+    	if(titulo == "" || descricao == "") {
+    		
+    		this.msgAlert.getMessageCampoEmBranco();
+    		
+    	}
+    	
+    	else {
+    		
+    		
+        	temp.setTitulo(titulo);
+        	temp.setDescricao(descricao);
+        	
+        	ProjetosScreenController.setProjetoSalvo(temp);
+        	
+        	this.msgAlert.getMessageProjetoSalvo();
+        	
+        	
+    	}
+    	
+    	
     }
   
     @Override
