+++ src/controller/FormularioTarefaScreenController.java
@@ -3,14 +3,13 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.miprogramacao.gerenciadordetarefas.controller;
+package controller;
 
 import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
 import Exceptions.ArgumentoInvalidoException;
-import Exceptions.ObjetoInexistenteException;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -20,13 +19,15 @@ import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuBar;
+import javafx.scene.control.MenuButton;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.Button;
+import javafx.scene.control.DatePicker;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
-import com.miprogramacao.gerenciadordetarefas.model.*;
 
 
 /**
@@ -34,45 +35,26 @@ import com.miprogramacao.gerenciadordetarefas.model.*;
  * @author User
  */
 
-public class FormularioProjetoScreenEditController implements Initializable {
+public class FormularioTarefaScreenController implements Initializable {
 	
-    @FXML
+	@FXML
     private TextField txtTitulo;
 
     @FXML
     private TextArea txtDescricao;
-    
-    private MessageAlert msgAlert = new MessageAlert();
-    
-    private Projeto temp = new Projeto("temp");
-    private Projeto projetoSelecionado = ProjetosScreenController.getProjetoSelecionado();
 
     @FXML
-    void addAlteracao(ActionEvent event) throws ArgumentoInvalidoException, ObjetoInexistenteException {
-    	
-    	String titulo = txtTitulo.getText();
-    	String descricao = txtDescricao.getText();
-    	
-    	if(titulo == "" || descricao == "") {
-    		
-    		this.msgAlert.getMessageCampoEmBranco();
-    		
-    	}
-    	
-    	else {
-    		
-    		
-        	temp.setTitulo(titulo);
-        	temp.setDescricao(descricao);
-        	
-        	ProjetosScreenController.setProjetoEditado(temp, projetoSelecionado.getTitulo());
-        	
-        	this.msgAlert.getMessageProjetoEditado();;
-        	
-        	
-    	}
-    	
-    	
+    private DatePicker txtValidade;
+
+    @FXML
+    private MenuButton statusMenuBTN;
+
+    @FXML
+    private Button addNovaTarefa;
+
+    @FXML
+    void salvarNovaTarefa(ActionEvent event) {
+
     }
   
     @Override
