+++ src/controller/TarefasScreenController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.miprogramacao.gerenciadordetarefas.controller;
+package controller;
 
 import java.io.IOException;
 import java.net.URL;
